requirements.txt
Django>=4.2,<5.0
 djangorestframework>=3.14,<4.0
 psycopg2-binary>=2.9,<3.0
 celery[redis]>=5.3,<6.0
 redis>=4.5,<5.0 
django-celery-results>=2.5,<3.0 
django-redis>=5.4,<6.0 
openai>=1.0.0
eventlet
pytest
pytest-django
pytest-cov
drf-yasg
python-dotenv